chore(components): remove aiplatform.batch_predict_job
PiperOrigin-RevId: 526769412
