+++ components/google-cloud/tests/v1/test_components_compile.py
@@ -50,6 +50,7 @@ from google_cloud_pipeline_components.v1.model import (
     ModelUploadOp,
     ModelExportOp,
 )
+from google_cloud_pipeline_components.v1.batch_predict_job import ModelBatchPredictOp
 from kfp import compiler
 from kfp.dsl import Input, Artifact
 
@@ -432,3 +433,48 @@ class ComponentsCompileTest(unittest.TestCase):
     del executor_output_json["sdkVersion"]
     del executor_output_json["schemaVersion"]
     self.assertEqual(executor_output_json, expected_executor_output_json)
+
+  def test_batch_prediction_op_compile(self):
+    @kfp.dsl.pipeline(name="training-test")
+    def pipeline(unmanaged_container_model: Input[Artifact]):
+      model_upload_op = ModelUploadOp(
+          project=self._project,
+          display_name=self._display_name,
+          unmanaged_container_model=unmanaged_container_model,
+      )
+
+      batch_predict_op = ModelBatchPredictOp(
+          project=self._project,
+          location=self._location,
+          job_display_name=self._display_name,
+          model=model_upload_op.outputs["model"],
+          instances_format="instance_format",
+          gcs_source_uris=[self._gcs_source],
+          bigquery_source_input_uri="bigquery_source_input_uri",
+          model_parameters={"foo": "bar"},
+          predictions_format="predictions_format",
+          gcs_destination_output_uri_prefix=self._gcs_destination_prefix,
+          bigquery_destination_output_uri="bigquery_destination_output_uri",
+          machine_type="machine_type",
+          accelerator_type="accelerator_type",
+          accelerator_count=1,
+          starting_replica_count=2,
+          max_replica_count=3,
+          manual_batch_tuning_parameters_batch_size=4,
+          generate_explanation=True,
+          explanation_metadata={"xai_m": "bar"},
+          explanation_parameters={"xai_p": "foo"},
+          encryption_spec_key_name="some encryption_spec_key_name",
+          labels={"foo": "bar"},
+      )
+
+    compiler.Compiler().compile(
+        pipeline_func=pipeline, package_path=self._package_path
+    )
+
+    with open(self._package_path) as f:
+      executor_output_json = json.load(f, strict=False)
+    with open("testdata/batch_prediction_pipeline.json") as ef:
+      expected_executor_output_json = json.load(ef, strict=False)
+    # Ignore the kfp SDK version during comparison
+    del executor_output_json["sdkVersion"]
